Create tasks002.html
Making the descriptions expandable to make the page look cleaner. This is a classic "accordion" or "toggle" feature, and we add it with a few small changes to the tasks.html file.

The best way to do this is to make the entire task title clickable. When clicked, it will reveal the description below it.

Updated tasks.html with Expandable Descriptions