Update tasks004.html
A  bug crept into the JavaScript that would prevent all the functions from working. The issue is a technical one where the command getElementsByClassName was used incorrectly, causing a script error that stopped everything.

The fix is simple: we just need to use the modern and correct command, querySelectorAll, in its place. Here is the fully corrected code that should restore all functionality.

Corrected tasks.html File

This version corrects the bug in both the sortTasks and filterTasks functions. We replace the current code with this.